Refactor: `DTO`를 각 도메인에 알맞은 디렉터리로 이동
