+++ src/main/java/com/team7/retriever/domain/channel/controller/dto/response/ChannelDataResponse.java
@@ -1,0 +1,13 @@
+package com.team7.retriever.domain.channel.controller.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+public class ChannelDataResponse {
+	private String message;
+	private String status;
+}
