+++ src/main/java/com/team7/retriever/domain/crawling/controller/dto/response/SerpApiCrawlingResponse.java
@@ -1,0 +1,25 @@
+package com.team7.retriever.domain.crawling.controller.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@Builder
+public class SerpApiCrawlingResponse {
+
+	private List<CrawlRes> google;
+	private List<String> telegrams;
+
+	@Getter
+	@Setter
+	@Builder
+	public static class CrawlRes {
+		private String link;
+		private String title;
+		private String source;
+	}
+}
