--- src/main/java/com/team7/retriever/dto/ChDataResponse.java
@@ -1,14 +1,0 @@
-package com.team7.retriever.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class ChDataResponse {
-    private String message;
-    private String status;
-
-}
