--- src/main/java/com/team7/retriever/dto/ChatArgotDrugDTO.java
@@ -1,19 +1,0 @@
-package com.team7.retriever.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ChatArgotDrugDTO {
-
-    private int id;
-    private List<String> argot;
-    private List<String> drugs;
-    private LocalDateTime timestamp;
-}
