--- src/main/java/com/team7/retriever/dto/UpdateCheckRequest.java
@@ -1,3 +1,0 @@
-package com.team7.retriever.dto;
-
-public record UpdateCheckRequest(String link, String title, String source) {}
