--- src/main/java/com/team7/retriever/dto/WebCrawlingResponse.java
@@ -1,12 +1,0 @@
-package com.team7.retriever.dto;
-
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-public class WebCrawlingResponse {
-    // private List<CrawlGoogleResponse> google;
-    private List<String> google;
-    private List<String> telegrams;
-}
